\name{Bermuda-package}
\alias{Bermuda-package}
\alias{Bermuda}
\docType{package}
\title{
\packageTitle{Bermuda}
}
\description{
\packageDescription{Bermuda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bermuda}
\packageIndices{Bermuda}
Provide Bernoulli and multinomial distribution allocation.
}
\author{
\packageAuthor{Bermuda}

Maintainer: \packageMaintainer{Bermuda}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
