\name{Bermuda-package}
\alias{Bermuda-package}
\alias{Bermuda}
\docType{package}
\title{
\packageTitle{Bermuda}
}
\description{
\packageDescription{Bermuda}
}
\details{
Provide Bernoulli and multinomial distribution allocation.
}
\author{
\packageAuthor{Bermuda}

Maintainer: \packageMaintainer{Bermuda}
}
