\name{BERMUDA-package}
\alias{BERMUDA-package}
\alias{BERMUDA}
\docType{package}
\title{
\packageTitle{BERMUDA}
}
\description{
\packageDescription{BERMUDA}
}
\details{
Provide Bernoulli and multinomial distribution allocation.
}
\author{
\packageAuthor{BERMUDA}

Maintainer: \packageMaintainer{BERMUDA}
}
